fix: A3 — move FuncFilter into internal/funkutil (shared with the shim)
FuncFilter/NewFuncFilter/Match/Sidecar lived in cmd/enumerate.go, so the
shim (a separate package main, cmd/shim_binary) couldn't import them and
had to hand-roll its own includeRe/excludeRe fields plus a Tracer.matchesFilter
that duplicated FuncFilter.Match exactly, and re-compiled the sidecar's
regex patterns inline in NewTracer instead of sharing one code path.

Move the type into internal/funkutil (new filter.go, alongside the
FilterSidecar it already owns), update every cmd-package call site to
funkutil.FuncFilter/funkutil.NewFuncFilter, and add
funkutil.FilterFromSidecar for the shim. Tracer gains a single
*funkutil.FuncFilter field replacing includeRe/excludeRe; matchesFilter is
gone, replaced by filter.Match directly (nil-safe by construction, same as
before).

Lenient-on-bad-pattern behavior is preserved but the diagnostic changes
slightly: FilterFromSidecar silently drops a malformed pattern rather than
returning the compile error, so NewTracer's debugLog message names the bad
pattern but no longer includes regexp's own error text.

FuncFilter's own unit tests (TestNewFuncFilter, TestFuncFilterMatch,
TestFuncFilterSidecar) moved to internal/funkutil/filter_test.go with the
type; added TestFilterFromSidecar. All other existing tests pass with only
mechanical NewFuncFilter -> funkutil.NewFuncFilter call-site updates.
