+++ cmd/trace.go
@@ -14,7 +14,7 @@ import (
 // Writes _functions.log, creates a temporary real-binary symlink in
 // SAFE_BIN_DIR so the shim's path-convention lookup works, invokes the
 // shim, then cleans up.
-func traceInline(binaryPath string, args []string, noLibs bool, filter *FuncFilter) (int, error) {
+func traceInline(binaryPath string, args []string, noLibs bool, filter *funkutil.FuncFilter) (int, error) {
 	logDir := funkutil.LogDir()
 	safeBinDir := funkutil.SafeBinDir()
